requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-dotenv==0.19.0
pyjwt==2.1.0
flask-bcrypt==0.7.1
flask-restful==0.3.9
marshmallow==3.13.0